hello

+++ src/main/webapp/WEB-INF/views/viewcategory.jsp
@@ -14,6 +14,26 @@
 
 <!-- Latest compiled JavaScript -->
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+
+
+
+<script type="text/javascript">
+	
+	$(document).ready(function() {
+		
+		$('#checkBoxAll').click(function() {
+			
+			if($(this).is(":checked"))
+				$('.chkCheckBoxId').prop('checked', true);
+			else
+				$('.chkCheckBoxId').prop('checked', false);
+		});
+	});
+
+</script>
+
+
+
 </head>
 <body>
 
@@ -22,12 +42,21 @@
 
 	<h1 style="margin: auto; width: 80%; text-align: center;">View Category</h1>
 
+
+	<form method="post" action="${pageContext.request.contextPath}/DeleteCategories">
+	
+	<input type="submit" class="btn btn-danger" value="Delete" onclick="return confirm('Are you sure you want to delete this category?');" style="margin-left: 135px;">	
+	
+	<br>
+	<br>
+
 	<br>
 	
 	<table class="table table-striped" style="margin: auto; width: 80%;">
 	
 		<thead>
 			<tr>
+			<th><input type="checkbox" id="checkBoxAll"></th> 
 			<th>Name</th>
 			<th>Description</th>
 			<th>View</th>
@@ -39,11 +68,12 @@
 		<tbody>
 			<c:forEach items="${categorydata}" var="x">
 			<tr>
+			<td><input type="checkbox" class="chkCheckBoxId" name="delete" value="${x.getCatid()}"></td>
 			<td>${x.getName()}</td>
 			<td>${x.getDescription()}</td>
 			<td><a href="ViewOneCategory/${x.getCatid()}" class="btn btn-primary">View</a></td>
 			<td><a href="UpdateOneCategory/${x.getCatid()}" class="btn btn-success">Update</a></td>
-			<td><a href="DeleteOneCategory/${x.getCatid()}" class="btn btn-danger">Delete</a></td>
+			<td><a href="DeleteOneCategory/${x.getCatid()}" class="btn btn-danger" onclick="return confirm('Are you sure you want to delete this category?');">Delete</a></td>
 			</tr>
 			</c:forEach>
 		</tbody>
